fixes for moving NCO calls into netcdf_utils, clean up
